requirements.txt
transformers==4.41.2
streamlit==1.35.0
torch==2.3.1
pillow==10.3.0
bitsandbytes~=0.43.1
accelerate~=0.31.0